Ok here we go, deals with a bunch of badness that can happen - Fixes a bug with displaying mounts in a grid (was probably due to nested for loops) - The main point being do requests only when you need to, keeping a local copy saves some data transfer.

+++ BlizzAPI.py
@@ -53,37 +53,20 @@ def main():
     root = Tk()
     root.title("Blizz API")
 
-    # response = requests.get(build_request(), params=get_params())
-
-    # Hey look, requests does json!
-    # print(response.json())
-
-    # We can even get specific things from it!
-    #  level = response.json()['level']
-    #  name = response.json()['name']
-    #  print(str(name) + ", level: " + str(level))
-
-    #  thumbnail_url = "http://render-us.worldofwarcraft.com/character/" + \
-    #                 str(response.json()['thumbnail'])
-    # profile_url = thumbnail_url[:-10] + 'profilemain.jpg'
-    #  inset_url = thumbnail_url[:-10] + 'inset.jpg'
     mounturl = "http://media.blizzard.com/wow/icons/36/"
     mountlisturl = "https://us.api.battle.net/wow/mount/?locale=en_US"
-#    mounts = requests.get(mountlisturl,
-#                          params={'apikey': get_api_key_from_file()})
-#    data = mounts.json()
-
     data = ""
+    imgs = []
 
-    try:
-        with open('../mounts.json', "r") as f:
+    p = Path('../mounts.json')
+
+    if p.exists():
+        with open(p, "r") as f:
             data = json.loads(f.read())
-    except FileNotFoundError:
-        with open('../mounts.json', "w") as f:
+    else:
+        with open(p, "w") as f:
             data = requests.get(mountlisturl, params={'apikey': get_api_key_from_file()}).json()
             json.dump(data, f)
-
-    image_bytes = []
 
     p = Path('../imgs/')
 
@@ -91,91 +74,35 @@ def main():
         p.mkdir(parents=True, exist_ok=True)
 
     for mount in data['mounts']:
-
-        try:
-            with open('../imgs/' + str(mount['creatureId']) + '.jpg', "rb") as f:
-                image_bytes.append(f.read())
-        except FileNotFoundError:
-            with open('../imgs/' + str(mount['creatureId']) + '.jpg', "wb") as f:
+        p = Path('../imgs/' + str(mount['creatureId']) + '.jpg')
+        if p.exists():
+            try:
+                with Image.open(p, "r") as f:
+                    imgs.append(ImageTk.PhotoImage(f))
+            except OSError:
+                p.unlink()
+                continue
+        else:
+            with open(p, "wb") as f:
                 response = requests.get(mounturl + mount['icon'] + '.jpg')
-                image_bytes.append(response.content)
-                f.write(response.content)
 
-    imgs = []
+                if response.status_code == 404:  # No mount image found
+                    continue  # So skip it.
+                else:
+                    imgs.append(ImageTk.PhotoImage(Image.open(BytesIO(response.content))))
+                    f.write(response.content)
 
-    for i in image_bytes:
-
-        imgs.append(ImageTk.PhotoImage(Image.open(BytesIO(i))))
-'''
-
-    imgs = []
     labels = []
 
-    for r in responses:
+    for i in range(len(imgs)):
+        labels.append(Label(root, image=imgs[i]).grid(row=i // 40, column=i % 40))
         try:
-            imgs.append(ImageTk.PhotoImage(Image.open(BytesIO(r.content))))
-        except OSError:
+            labels[i].pack(anchor=NW)
+        except AttributeError:
             continue
 
-    for i in range(25):
-        for j in range(47):
-            try:
-                labels.append(Label(root, image=imgs[i*25+j]).grid(row=i, column=j))
-                labels[i*25+j].pack(anchor=NW)
-            except AttributeError:
-                continue
-            except IndexError:
-                pass
     root.mainloop()
-'''
 
-'''
-    label = []
-    for i in range(10):
-        new_url = mounturl + data['mounts'][i]['icon'] + '.jpg'
-        r = requests.get(new_url)
-        img = ImageTk.PhotoImage(Image.open(BytesIO(r.content)))
-        label.append(Label(root, image=img))
-    root.mainloop()
-'''
-
-'''
-    for x in data['mounts']:
-        label = []
-        try:
-            new_url = mounturl + x['icon'] + '.jpg'
-            r = requests.get(new_url)
-            img = ImageTk.PhotoImage(Image.open(BytesIO(r.content)))
-            label.append(Label(root, image=img))
-            label.pop().pack()
-        except OSError:
-            print(new_url)
-            continue
-'''
-
-#   https://us.api.battle.net/wow/mount/ to get
-# mount list (takes locale and api key as params)
-#  avatar = requests.get(thumbnail_url)
-#  profile = requests.get(profile_url)
-#  inset = requests.get(inset_url)
-#  maini = requests.get((thumbnail_url[:-10] + 'main.jpg'))
-#  mount = requests.get()
-
-#   img = ImageTk.PhotoImage(Image.open(BytesIO(avatar.content)))
-#  pimg = ImageTk.PhotoImage(Image.open(BytesIO(profile.content)))
-#   iimg = ImageTk.PhotoImage(Image.open(BytesIO(inset.content)))
-#   mimg = ImageTk.PhotoImage(Image.open(BytesIO(maini.content)))
-# mountimg =
-
-#   label2 = Label(root, image=mimg)
-#  label2.pack()
-
-'''
-    label = Label(root, image=img)
-    label.pack()
-    label1 = Label(root, image=pimg)
-    label1.pack()
-'''
 
 # Just in case you use this somewhere else?
 if __name__ == "__main__":
